+++ app/models/peep.rb
