+++ app/views/layout.erb
@@ -19,6 +19,10 @@
   </form>
   <% end %>
 
+  <% if flash[:notice] %>
+    <div id='notice'><%= flash[:notice] %></div>
+  <% end %>
+
 <%= yield %>
 </body>
 </html>
