+++ spec/spec_helper.rb
@@ -2,14 +2,14 @@
 
 ENV['RACK_ENV'] = 'test'
 
-require File.join(File.dirname(__FILE__), '..', 'app.rb')
+require File.join(File.dirname(__FILE__), '..', 'app/app.rb')
 
 require 'capybara'
 require 'capybara/rspec'
 require 'rspec'
 require 'coveralls'
 require 'simplecov'
-require './models/user.rb'
+require './app/models/user.rb'
 require 'database_cleaner'
 
 SimpleCov.formatters = [
